implement permutations and bar re-ordering
git-svn-id: svn://svn.rforge.net/Acinonyx/trunk@25 5d8255c7-83f2-4cbe-a1b0-67fca80f6e0e

+++ Acinonyx/ATypes.h
@@ -17,7 +17,7 @@ extern "C" {
 	
 	/* basic types */
 
-	typedef float AFloat;
+	typedef double AFloat;
 	
 	typedef unsigned int vsize_t;
